requirements.txt
# Core deep learning (install one or both depending on your model choice)
# TensorFlow / Keras – required for VGG-Face via DeepFace
tensorflow>=2.12.0
keras>=2.12.0
# PyTorch – required for FaceNet (facenet-pytorch) and the custom CNN
torch>=2.0.0
torchvision>=0.15.0
deepface>=0.0.79
facenet-pytorch>=2.5.3
opencv-python>=4.7.0
dlib>=19.24.0
Pillow>=9.5.0

# Data handling and analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deployment
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0

# MLOps and monitoring
mlflow>=2.4.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
